MODE-1714 - Moved PageKey to the "spi" package
