+++ modeshape-jcr/src/main/java/org/modeshape/jcr/federation/spi/Pageable.java
@@ -25,7 +25,6 @@
 package org.modeshape.jcr.federation.spi;
 
 import org.infinispan.schematic.document.Document;
-import org.modeshape.jcr.federation.PageKey;
 
 /**
  * Marker interface that should be implemented by {@link org.modeshape.jcr.federation.spi.Connector}(s) that want to expose children
@@ -40,7 +39,7 @@ public interface Pageable {
      * Returns a document which represents a page of children. In order to add a next page, {@link PagingWriter#addPage(String, String, long, long)}
      * should be used to add a new page of children.
      *
-     * @param pageKey a {@code non-null} {@link org.modeshape.jcr.federation.PageKey} instance, which offers information
+     * @param pageKey a {@code non-null} {@link PageKey} instance, which offers information
      * about the page that should be retrieved.
      * @return either a {@code non-null} page document or {@code null} indicating that such a page doesn't exist
      */
